Markers are parsing!!
At this stage, it is largely complete now.  Remaining todo items:

(1) More multi-frame tests.
(2) Mode invalid tests.
(3) Real-world tests.

In the latter, we want to test against the official language suite to
make sure we can parse it all.

Having done all of that, the next step is to figure out how to apply a
testfile to produce a snapshot of the state at each point.  There is
probably a fair bit of work to do there.